requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1
jinja2>=3.1.3

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Cache / Queue
redis>=5.0.1
celery>=5.3.6
cachetools>=5.3.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI / ML
google-genai>=1.0.0
qdrant-client>=1.7.0
open-clip-torch>=2.24.0

# Computer Vision
opencv-python-headless>=4.9.0
ultralytics>=8.1.0
numpy>=1.26.0

# Config / Logging
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
structlog>=24.1.0

# Reports
python-docx>=1.1.0

# MQTT / IoT
aiomqtt>=2.0.0

# SSO / Identity
authlib>=1.3.0
pyotp>=2.9.0

# Observability
prometheus-client>=0.20.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp>=1.22.0

# Utilities
httpx>=0.26.0
pillow>=10.2.0
